User can now set missing Transaction:Category direclty from any Transaction List (#51)
* fix Account.fieldOpeningBalance serialize as Double

* User can now set missing Transaction:Category direclty from any Transaction List

---------

+++ lib/app/data/models/money_objects/accounts/account.dart
@@ -243,7 +243,8 @@ class Account extends MoneyObject {
     name: 'Opening Balance',
     serializeName: 'OpeningBalance',
     getValueForDisplay: (final MoneyObject instance) => (instance as Account).fieldOpeningBalance.value,
-    getValueForSerialization: (final MoneyObject instance) => (instance as Account).fieldOpeningBalance.value,
+    getValueForSerialization: (final MoneyObject instance) =>
+        (instance as Account).fieldOpeningBalance.value.toDouble(),
   );
 
   /// ReconcileWarning
